refactor(count): Change found_rows naming to count, #31
